backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
sqlmodel==0.0.22
asyncpg==0.29.0
redis==5.0.8
clickhouse-connect==0.7.19
kafka-python==2.0.2
pika==1.3.2
langchain==0.2.14
langchain-openai==0.1.23
python-dotenv==1.0.1
livekit==1.0.23
livekit-api==1.1.0
